+++ src/main/java/ru/itmo/javazolotaya/lesson1/SmenilNazvaniyeJavaZolotaya.java
@@ -1,13 +1,13 @@
-package ru.itmo.javazolotaya;
+package ru.itmo.javazolotaya.lesson1;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class JavazolotayaApplication {
+public class SmenilNazvaniyeJavaZolotaya {
 
 	public static void main(String[] args) {
-		SpringApplication.run(JavazolotayaApplication.class, args);
+		SpringApplication.run(SmenilNazvaniyeJavaZolotaya.class, args);
 	}
 
 }
