+++ src/main/java/ru/itmo/javazolotaya/lesson1/dto/DtoExample.java
@@ -1,4 +1,4 @@
-package ru.itmo.javazolotaya.dto;
+package ru.itmo.javazolotaya.lesson1.dto;
 
 import lombok.Data;
