+++ src/test/java/ru/itmo/javazolotaya/SmenilNazvaniyeJavaZolotayaTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class JavazolotayaApplicationTests {
+class SmenilNazvaniyeJavaZolotayaTests {
 
 	@Test
 	void contextLoads() {
